backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
qdrant-client
google-genai
pypdf
python-multipart
python-jose[cryptography]
email-validator
